Completed. start form

+++ database/migrations/2021_05_20_113537_create_comics_table.php
@@ -18,7 +18,7 @@ class CreateComicsTable extends Migration
             $table->string("title");
             $table->text("description");
             $table->text("thumb");
-            $table->tinyInteger("price");
+            $table->float("price");
             $table->string("series");
             $table->date("sale_date");
             $table->string("type");
